backend/requirements.txt
fastapi
uvicorn
vllm
mlx-vlm
pillow
huggingface_hub
openai
